feat: add 1.17 and 1.18

+++ example/testcase/assert_test.go
@@ -9,7 +9,7 @@ import (
 
 func TestAssertInt(t *testing.T) {
 	var a int64 = 1
-	var b int = 1
+	var b int8 = 1
 	la.Equal(t, a, b)
 	assert.NotEqual(t, a, b)
 }
